Use hash in Find op
Summary:
The logic for using the key hash in `FetchSEPrivKeyRef` and  `DeleteKey` is inverted. If the hash is not `NULL`, we should use it when fetching keys from SE.

It also looks like there is a longstanding bug in our assumption that the hash can be `NULL`. When we pass a `nil` hash in the Go code, because the zero value of a slice is not `nil`, memory is being allocated. Hence, even a 0 length []byte seems not to resolve to `NULL`. I believe this is what led to the inverted hash check in `macos.c`.

This changes the logic to only use the []byte hash if it contains any bytes.

Also, surface some additional errors. In a later PR, we can have some per-OS messages that include the hash.

Reviewed By: n-canter

Differential Revision: D34001233

fbshipit-source-id: a62c0bc266a53eb4a260d52cfb73ac1b162c5123