Add LoadKey method
Summary: This diff separates the logic a bit around creating a new key and loading an existing key. This is because if we pass a hash to `NewKey` and that key isn't found, we will create a new key. So, if that key for that hash we lookup gets accidentally deleted (only pointing fingers at myself here), we will then continuously create new keys.

Reviewed By: n-canter

Differential Revision: D34304696

fbshipit-source-id: 2529fce5b807d387b57b4f4493e1728585ace539