Device-file check alteration
Summary:
== Summary
ModeSocket should also be included into the mode check.

== Important notes
If the check was to fail, the agent should have been broken, after the start up, but it works, which indicates that the check passes. The interface also is being tested and agent succeeds with requests to swtpm through the socket.

Go tpm2 docs say that the method OpenTPM can accept either a chardev file, or a socket
== What/Why?
This is necessary to run the agent with swtpm for the E2E testing project.

Reviewed By: n-canter

Differential Revision: D42843160

fbshipit-source-id: 4e4fef765a5b448742f26badf76e3aa98a8694f2