Handle initialised system with no persistent keys
Summary:
The `nextAvailable` function that looks for available persistent TPM
handles can cope when there are no persistent handles, but not when
there are some persistent handles but none of them are in our allocation
range. In this instance we filter them all out, end up with an empty
slice and promptly try to do something with element 0. Which crashes.

Check if we end up with no handles left after filtering, and if so
return the first handle as we do in the completely empty case.

Reviewed By: shapzka

Differential Revision: D52628849

fbshipit-source-id: b580bfd7ffecf752f17dd8851ad72022912e9d48